% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{AIC,Dlm-method}
\alias{AIC,Dlm-method}
\title{Akaike Information Criterion}
\usage{
\S4method{AIC}{Dlm}(object, ..., k = 2)
}
\value{
A numeric scalar: AIC
}
\description{
Compute AIC statistic for fitted DLM
}
