% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{BayesDlm}
\alias{BayesDlm}
\alias{.BayesDlm}
\title{Bayesian Distributed Lag Model}
\description{
Parameters and summaries from a Bayesian DLM. Inherits from
\code{\link{Dlm}}.
}
\section{Slots}{

\describe{
\item{\code{fitted}}{fitted values; conditional expectation of observed \code{y}}

\item{\code{DIC}}{Deviance Information Criterion}

\item{\code{residuals}}{observed residuals}

\item{\code{sims}}{list of posterior simulation draws of model parameters}
}}

\references{
Baek J, et al. (2016) Epidemiology 27(1):116-24.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/26414942}{PubMed})
}
\seealso{
\code{\link{dlm}}
}
