% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{BIC,Dlm-method}
\alias{BIC,Dlm-method}
\title{"Bayesian" Information Criterion}
\usage{
\S4method{BIC}{Dlm}(object, ...)
}
\value{
A numeric scalar: BIC
}
\description{
Compute BIC statistic for fitted DLM
}
