% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{coef,Dlm-method}
\alias{coef,Dlm-method}
\title{Extract Distributed Lag Model Coefficients}
\usage{
\S4method{coef}{Dlm}(object)
}
\value{
A numeric vector of fitted coefficients
}
\description{
Return DLM coefficients
}
