% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{coef.Dlm}
\alias{coef.Dlm}
\title{Extract Distributed Lag Model Coefficients}
\usage{
\method{coef}{Dlm}(object, ...)
}
\value{
A numeric vector of fitted coefficients.
}
\description{
Return DLM coefficients
}
\details{
Note that these are the basis-scaled coefficients. To extract
  coefficients on the scale of the original design (\eqn{\theta}),
  see \code{\link{theta}}.
}
