% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{coef.sims,BayesDlm-method}
\alias{coef.sims,BayesDlm-method}
\title{Extract Distributed Lag Coefficient Simulation Draws}
\usage{
\S4method{coef.sims}{BayesDlm}(object, ...)
}
\value{
A numeric matrix containing the scaled simulated coefficient matrix
}
\description{
\code{Dlm} objects fit by the Bayesian method store a matrix of
simulated coefficients drawn via MCMC from their full conditional
distribution. Extract the distributed lag-scaled simulated coefficient
matrix.
}
