% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{coef.sims}
\alias{coef.sims}
\title{Extract Coefficient Simulations}
\usage{
coef.sims(object, ...)
}
\arguments{
\item{object}{a fitted model object}
}
\value{
A numeric matrix
}
\description{
Extract simulated coefficients from a model fitted with MCMC
}
\details{
\code{coef.sims} is S4 generic.
}
