% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{Dlm-class}
\alias{Dlm-class}
\alias{.Dlm}
\title{S4 Distributed Lag Models}
\description{
Parent class for DLM object representation. Users will interact mainly
with daughter classes, \code{\link{FreqDlm}} and \code{\link{BayesDlm}}.
}
\details{
Coming soon!
}
\section{Slots}{

\describe{
\item{\code{method}}{character identifier for method (frequentist vs. Bayesian)
used to fit the DLM}

\item{\code{call}}{language object containing matching function call}

\item{\code{coefficients}}{DLM coefficients}

\item{\code{sigma}}{residual standard deviation}

\item{\code{tau}}{random effects standard deviation}

\item{\code{logLik}}{log likelihood}

\item{\code{vcov}}{fitted variance-covariance matrix}

\item{\code{basis}}{\code{\link{LagBasis}} object containing lag basis representation}

\item{\code{N}}{Number of data points the model was fit to}

\item{\code{K}}{list of coefficient type lengths. See 'Details.'}
}}

\references{
Baek J, et al. (2016) Epidemiology 27(1):116-24.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/26414942}{PubMed})
}
\seealso{
\code{\link{dlm}}, \code{\link{LagBasis}}
}
