% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm.R
\name{dlm}
\alias{dlm}
\title{Distributed Lag Models}
\usage{
dlm(y, X, Z, lags, intercept = TRUE, method = c("frequentist", "bayesian",
  "Frequentist", "Bayesian"), control = list(), ...)
}
\arguments{
\item{y}{model response variable}

\item{X}{matrix of lag count data}

\item{Z}{covariate matrix  --> TODO: make optional?}

\item{lags}{lag distances for columns of \code{X}}

\item{intercept}{\code{c(TRUE, FALSE)}; \code{TRUE} if an intercept
should be included in the model}

\item{method}{\code{c("frequentist", "bayesian")}; method used to fit the
DLM. Partial matching and capitalization allowed}

\item{constrol}{\code{list(...)}; a list of control parameters.
See 'Details'}
}
\value{
An S4 \code{Dlm} object containing the results of the fitted model
}
\description{
Fit distributed lag models to distance-profiled data.
\code{frequentist} method relies on \code{nlme::lme}.
\code{bayesian} method uses Gibbs to sample from full conditionals,
with sampling parameters named in the \code{control} list.
}
\details{
\code{control} list specifies additional optional \code{"bayes"} method
arguments, and may include: \code{n.sims}, the total number of simulations
to run; \code{n.save}, the total number of simulations to save;
\code{burnin}, number of simulations to discard from the start of the chain;
\code{alpha.tau.prior}, prior (shape) parameter \eqn{\alpha_{\tau^2}};
\code{beta.tau.prior}, prior (rate) parameter \eqn{\beta_{\tau^2}};
\code{alpha.sigma.prior}, prior (shape) parameter \eqn{\alpha_{\sigma^2}};
\code{beta.sigma.prior}, prior (rate) parameter \eqn{\beta_{\sigma^2}}.

The prior distribution heirarchy we assume in the Bayesian
formulation of the DLM is as follows:

\deqn{
  y \sim N(D \theta, \tau^2 I_n) \\
  \theta \sim N(\mu_{\theta}, \Sigma_{\theta}) \\
  \theta_{(lag)} \sim N(\mu_{\theta_{(lag)}}, \sigma^2 I_L) \\
  \tau^2 \sim Inv-Gamma(\alpha_{\tau^2}, \frac{1}{\beta_{\tau^2}}) \\
  \sigma^2 \sim Inv-Gamma(\alpha_{\sigma^2}, \frac{1}{\beta_{\sigma^2}})
}
}
\examples{
data (simdata)
y <- simdata[, 1]
Z <- as.matrix(simdata[, 2:3])
X <- as.matrix(simdata[, 4:ncol(simdata)])
lg <- seq(0.1, 10, length.out = ncol(X))
fit <- dlm(y, X, Z, lg)
fit

}
\references{
Baek J, et al. (2016) Epidemiology 27(1):116-24.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/26414942}{PubMed})
}
