% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{Dlm_Summary}
\alias{Dlm_Summary}
\title{Summarize a \code{Dlm} Object}
\usage{
\S4method{summary}{FreqDlm}(object, level = 0.95, ...)
}
\value{
A '\code{SummaryDlm}' object with information about the fitted model.
}
\description{
Summarize a fitted DLM object. Details coming soon!
}
\seealso{
\code{\link{SummaryDlm}}
}
