% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{fitted,Dlm-method}
\alias{fitted,Dlm-method}
\title{Extract Model Fitted Values}
\usage{
\S4method{fitted}{Dlm}(object)
}
\value{
A numeric vector: the model fitted values
}
\description{
Will return predicted \code{y} values from a fitted DLM.
}
