% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{fitted.Dlm}
\alias{fitted.Dlm}
\title{Extract DLM Model Fitted Values}
\usage{
\method{fitted}{Dlm}(object)
}
\arguments{
\item{object}{a fitted \code{"\link[=Dlm-class]{Dlm}"} object}
}
\value{
A numeric vector
}
\description{
Extract the predicted \code{y} values from a fitted by \code{\link{dlm}}
}
