% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{fixef,Dlm-method}
\alias{fixef,Dlm-method}
\title{Extract Fixed Effects}
\usage{
\S4method{fixef}{Dlm}(object)
}
\value{
A numeric vector: fixed effect coefficients
}
\description{
Will return a vector of fixed effects coefficients from the fitted DLM.
}
