% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{fixef}
\alias{fixef}
\alias{fixef,Dlm-method}
\title{Extract Fixed Effects}
\usage{
fixef(object, ...)

\S4method{fixef}{Dlm}(object)
}
\value{
A numeric vector
}
\description{
Returns a vector of fixed effects coefficients from a fitted
mixed effects model.
}
\details{
\code{fixef} is S4 generic.
}
\section{Methods (by class)}{
\itemize{
\item \code{Dlm}: S4 Method for \code{"\link[=Dlm-class]{Dlm}"} Objects
}}

