% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{FreqDlm}
\alias{FreqDlm}
\alias{.FreqDlm}
\title{Classical Distributed Lag Model}
\description{
Parameters and summaries from a classically fit DLM (freqentist;
model fit using \code{\link{nlme::lme}}). Inherits from
\code{\link{Dlm}}.
}
\section{Slots}{

\describe{
\item{\code{lme}}{output of fitted \code{lme} object}
}}

\references{
Baek J, et al. (2016) Epidemiology 27(1):116-24.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/26414942}{PubMed})
}
\seealso{
\code{\link{dlm}}, \code{\link{lme}}
}
