% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{logLik,Dlm-method}
\alias{logLik,Dlm-method}
\title{Extract Log-Likelihood}
\usage{
\S4method{logLik}{Dlm}(object)
}
\value{
An object of class \code{logLik} containing the log-likelihood of the
data given the fitted DLM
}
\description{
Log likelihood of the observed data given the model
}
