% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{logLik.Dlm}
\alias{logLik.Dlm}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{Dlm}(object, ...)
}
\value{
An object of class \code{logLik} containing the log-likelihood of the
data given the fitted DLM
}
\description{
Log likelihood of the observed data given the model
}
