% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{omega,LagBasis-method}
\alias{omega,LagBasis-method}
\title{Extract Lag Basis Matrix}
\usage{
\S4method{omega}{LagBasis}(object, ...)
}
\value{
A square numeric basis matrix
}
\description{
Extract lag basis matrix, \eqn{\Omega}
}
