% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Dlm}
\alias{plot.Dlm}
\title{Plot a Fitted Dlm}
\usage{
\method{plot}{Dlm}(x, y, ...)
}
\arguments{
\item{x}{a fitted \code{"\link{Dlm-class}"} object}

\item{...}{additional arguments passed to \code{\link{summary.Dlm}}}
}
\value{
A \code{ggplot2} graphic object
}
\description{
Plots point estimates and intervals for distributed lag terms from a
fitted \code{\link{dlm}}. Default aesthetic relies on
\code{\link[ggplot2]{geom_pointrange}} from the \code{ggplot2} package.
}
