% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{predict,LagBasis-method}
\alias{predict,LagBasis-method}
\title{Predict New Values for Fitted Lag Basis}
\usage{
\S4method{predict}{LagBasis}(object, ...)
}
\description{
Not yet implemented
}
