% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{ranef,Dlm-method}
\alias{ranef,Dlm-method}
\title{Extract Random Distributed Lag Effects}
\usage{
\S4method{ranef}{Dlm}(object)
}
\value{
A numeric vector: random effect coefficients
}
\description{
Will return a vector of random effects coefficients
from the fitted DLM.
}
