% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{ranef}
\alias{ranef}
\alias{ranef,Dlm-method}
\title{Extract Random Effects}
\usage{
ranef(object, ...)

\S4method{ranef}{Dlm}(object)
}
\value{
A numeric vector
}
\description{
Returns a vector of random effects coefficients from a fitted
mixed effects model.
}
\details{
\code{ranef} is S4 generic.
}
\section{Methods (by class)}{
\itemize{
\item \code{Dlm}: S4 Method for \code{"\link[=Dlm-class]{Dlm}"} Objects
}}

