% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{residuals,Dlm-method}
\alias{residuals,Dlm-method}
\title{Extract Model Residuals}
\usage{
\S4method{residuals}{Dlm}(object)
}
\value{
A numeric vector: the model residuals
}
\description{
Will return the observed residuals from the model.
}
