% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{residuals.Dlm}
\alias{residuals.Dlm}
\title{Extract DLM Model Residuals}
\usage{
\method{residuals}{Dlm}(object)
}
\arguments{
\item{object}{a fitted \code{"\link[=Dlm-class]{Dlm}"} object}
}
\value{
A numeric vector
}
\description{
Extract residuals from a model fitted by \code{\link{dlm}}
}
