% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{scaleMat,Dlm-method}
\alias{scaleMat,Dlm-method}
\title{Extract Lag Coefficient Scale Matrix}
\usage{
\S4method{scaleMat}{Dlm}(object, ...)
}
\value{
A numeric matrix: the block diagonal scale matrix, S
}
\description{
Return lag coefficient scale matrix, S, such that the distributed lag
coefficients fit by the model are obtained via the transformation
\eqn{\beta = S \theta}.
}
