% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{scaleMat}
\alias{scaleMat}
\alias{scaleMat,Dlm-method}
\title{Extract Distributed Lag Scale Matrix}
\usage{
scaleMat(object, ...)

\S4method{scaleMat}{Dlm}(object, ...)
}
\arguments{
\item{object}{a fitted model object}
}
\value{
A square numeric matrix
}
\description{
Return lag coefficient scale matrix, S, such that the distributed lag
coefficients fit by the model are obtained via the transformation
\eqn{\beta = S \theta}{\beta = S * \theta}. S should be invertable.
}
\details{
\code{scaleMat} is S4 generic.
}
\section{Methods (by class)}{
\itemize{
\item \code{Dlm}: S4 Method for \code{"\link[=Dlm-class]{Dlm}"} Objects
}}

