% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{se.fixef,Dlm-method}
\alias{se.fixef,Dlm-method}
\title{Extract Fixed Effect Standard Errors}
\usage{
\S4method{se.fixef}{Dlm}(object)
}
\value{
A numeric vector: fixed effects coefficients standard errors
}
\description{
Will return a vector of standard errors for each fixed effect coefficient.
}
