% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{se.fixef}
\alias{se.fixef}
\alias{se.fixef,Dlm-method}
\title{Extract Standard Errors of Fixed Effects}
\usage{
se.fixef(object)

\S4method{se.fixef}{Dlm}(object)
}
\value{
A numeric vector
}
\description{
Returns a vector of fixed effects standard errors from a fitted
mixed effects model.
}
\details{
\code{fixef} is S4 generic.
}
\section{Methods (by class)}{
\itemize{
\item \code{Dlm}: S4 Method for \code{"\link[=Dlm-class]{Dlm}"} Objects
}}

