% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{se.ranef,Dlm-method}
\alias{se.ranef,Dlm-method}
\title{Extract Random Effect (Distributed Lag) Standard Errors}
\usage{
\S4method{se.ranef}{Dlm}(object)
}
\value{
A numeric scalar: the random effects (distributed lag) coefficient
standard error
}
\description{
Will return the random effects standard error. Note that this is a
single value in the non-heirarchical formulation of this model.
}
