% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{se.ranef}
\alias{se.ranef}
\alias{se.ranef,Dlm-method}
\title{Extract Standard Errors of Random Effects}
\usage{
se.ranef(object)

\S4method{se.ranef}{Dlm}(object)
}
\value{
A numeric vector
}
\description{
Returns a vector of random effects standard errors from a fitted
mixed effects model.
}
\details{
\code{fixef} is S4 generic.
}
\section{Methods (by class)}{
\itemize{
\item \code{Dlm}: S4 Method for \code{"\link[=Dlm-class]{Dlm}"} Objects
}}

