% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{show,Dlm-method}
\alias{show,Dlm-method}
\title{Display Results of a Fitted DLM}
\usage{
\S4method{show}{Dlm}(object)
}
\description{
Print basic output from a fitted DLM object. Output structured after
\code{nlme::lme.print()}
}
