% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{sigma,Dlm-method}
\alias{sigma,Dlm-method}
\title{Extract Residual Standard Deviation 'Sigma'}
\usage{
\S4method{sigma}{Dlm}(object)
}
\value{
A positive numeric scalar: the residual standard deviation
}
\description{
Return observed residual standard deviation
}
