% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{sigma.Dlm}
\alias{sigma.Dlm}
\title{Extract Residual Standard Deviation 'Sigma'}
\usage{
\method{sigma}{Dlm}(object, ...)
}
\value{
A positive numeric scalar: the residual standard deviation
}
\description{
Return observed residual standard deviation
}
