% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata-data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated Built Environment Data}
\format{The first column (\code{Y}) is the outcome variable, covariates \code{Gender}
and \code{Age} come next, and are followed by 100 location description
variables.}
\usage{
data(simdata)
}
\description{
Simulated Built Environment Data
}
\examples{
data(simdata)
simdata[1:10, 1:5]

}
\references{
Baek J, et al. (2016) Epidemiology 27(1):116-24.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/26414942}{PubMed})
}
\keyword{datasets}
