% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{summary.Dlm}
\alias{summary.Dlm}
\alias{summary.FreqDlm}
\alias{summary.BayesDlm}
\title{Summarize a Fitted Dlm Object}
\usage{
\method{summary}{FreqDlm}(object, level = 0.95, ...)

\method{summary}{BayesDlm}(object, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted \code{"\link[=Dlm-class]{Dlm}"} object}

\item{level}{desired confidence coefficient (or credible interval) level for
summarizing distributed lag terms}

\item{...}{ignored}
}
\value{
An S4 object of class \code{"\link{SummaryDlm}"}
}
\description{
Computes likelihood based statistics, t-statistic summaries, and correlations
between fixed effects terms after the \code{nlme} package. Also computes
rough distributional summaries for distributed lag terms (estimates and
intervals, with confidence coefficient level determined by the \code{level}
parameter).
}
