% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{SummaryDlm}
\alias{SummaryDlm}
\alias{.SummaryDlm}
\title{Summarize a Distributed Lag Model}
\description{
Compute and store various statistical summaries of a fitted distributed
lag model object
}
\section{Slots}{

\describe{
\item{\code{method}}{Character vector with details of the model fitting procedure}

\item{\code{call}}{Matching call from \code{dlm}}

\item{\code{coefficients}}{Numeric vector of distributed lag coefficients}

\item{\code{likelihood.stats}}{Named numeric vector of likelihood based statistical summaries}

\item{\code{variances}}{Named numeric vector with estimaed random effects and residual variance
parameters}

\item{\code{tTable}}{t-statistic table summaries of fixed effects with (approximate)
significance}

\item{\code{lag.coefs}}{A matrix with rough distributional summaries of the fitted lag
coefficients, like estimates and confidence or posterior intervals.}

\item{\code{cor.fixef}}{Correlation matrix of fixed effects}

\item{\code{residuals}}{Numeric vector with the residuals from the fitted model}

\item{\code{N}}{Number of data points the model was fit to}

\item{\code{K}}{List of numbers of various covariate types}
}}

