% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{theta}
\alias{theta}
\alias{theta,Dlm-method}
\title{Extract Natural Scale Model Parameters 'Theta'}
\usage{
theta(object, ...)

\S4method{theta}{Dlm}(object, ...)
}
\arguments{
\item{object}{a fitted model object}
}
\value{
A numeric vector
}
\description{
Extract \eqn{\theta} coefficients from a fitted model
}
\details{
Returns the natural scale coefficients
\eqn{\theta = S^{-1} \beta}{\theta = S^-1 * \beta} given some invertable
scale matrix, S.

\code{theta} is S4 generic.
}
\section{Methods (by class)}{
\itemize{
\item \code{Dlm}: S4 Method for \code{"\link[=Dlm-class]{Dlm}"} Objects
}}

