% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{vcov,Dlm-method}
\alias{vcov,Dlm-method}
\title{Calculate the Distributed Lag Variance-Covariance Matrix for a
  Fitted Model Object}
\usage{
\S4method{vcov}{Dlm}(object)
}
\value{
A numeric matrix: the coefficient variance-covariance matrix
}
\description{
Returns the variance-covariance matrix of the fitted model coefficients
(including the random effects terms).
}
