% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{vcov.Dlm}
\alias{vcov.Dlm}
\title{Calculate the Distributed Lag Variance-Covariance Matrix for a
  Fitted Model Object}
\usage{
\method{vcov}{Dlm}(object, ...)
}
\value{
A numeric matrix: the coefficient variance-covariance matrix
}
\description{
Returns the variance-covariance matrix of the fitted model coefficients
(including the random effects terms).
}
