% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{vcovTheta}
\alias{vcovTheta}
\alias{vcovTheta,Dlm-method}
\title{Extract Natural Scale Coefficient Covariance Matrix}
\usage{
vcovTheta(object, ...)

\S4method{vcovTheta}{Dlm}(object, ...)
}
\arguments{
\item{object}{a fitted model object}
}
\value{
A square numeric matrix
}
\description{
Extract \eqn{var \theta} matrix from a fitted model
}
\details{
Returns the covariance matrix for the natural scale coefficients,
\eqn{\theta}.

\code{vcovTheta} is S4 generic.
}
\section{Methods (by class)}{
\itemize{
\item \code{Dlm}: S4 Method for \code{"\link[=Dlm-class]{Dlm}"} Objects
}}

